Fix : clova api 사용 최종 검증
